refactor(dashboard): raw escape-hatch routes + delete the if-chain
SSE /api/events, binary /api/image and POST /api/ingest move behind
rawRoute (verbatim bodies; the IngestStreamBus seam is untouched);
ServeIngestState gets its own module so live routes and serve lifecycle
share it without a cycle. handleDashboardRequest is now dispatch + a
three-line epilogue; server.ts drops from 1,249 to ~250 lines.
